unify Enclave Mocks
