+++ service/src/tests/mocks/mod.rs
@@ -20,5 +20,4 @@ pub mod direct_request_mock;
 pub mod enclave_api_mock;
 pub mod initialization_handler_mock;
 pub mod parentchain_api_mock;
-pub mod sidechain_api_mock;
 pub mod update_worker_peers_mock;
