+++ service/src/tests/parentchain_handler_test.rs
@@ -17,9 +17,7 @@
 
 use crate::{
 	parentchain_handler::{HandleParentchain, ParentchainHandler},
-	tests::mocks::{
-		parentchain_api_mock::ParentchainApiMock, sidechain_api_mock::SidechainApiMock,
-	},
+	tests::mocks::{enclave_api_mock::EnclaveMock, parentchain_api_mock::ParentchainApiMock},
 };
 use itp_node_api::api_client::ChainApi;
 use std::sync::Arc;
@@ -32,7 +30,7 @@ fn test_number_of_synced_blocks() {
 	let last_synced_block =
 		parentchain_api_mock.get_blocks(2, 2).unwrap().first().cloned().unwrap();
 
-	let enclave_api_mock = SidechainApiMock;
+	let enclave_api_mock = EnclaveMock;
 
 	let parentchain_handler =
 		ParentchainHandler::new(parentchain_api_mock, Arc::new(enclave_api_mock));
